Add unbindFrom and NonBindable interface to Bindable
This update introduces the function unbindFrom to the Bindable interface, permitting a better management of bindings by facilitating the removal of existing bindings.
Also a NonBindable sub-interface was added, for providing a placeholder interface to use when no binding action is desired. Now, an object can explicitly be set as NonBindable, enhancing code readability and safety.